[cri] Move cri from ws-sync to common-go for reuse in ws-manager-node
