+++ components/ws-sync/pkg/syncd/service.go
@@ -15,13 +15,13 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/gitpod-io/gitpod/common-go/cri"
 	"github.com/gitpod-io/gitpod/common-go/tracing"
 	csapi "github.com/gitpod-io/gitpod/content-service/api"
 	wsinit "github.com/gitpod-io/gitpod/content-service/pkg/initializer"
 	"github.com/gitpod-io/gitpod/content-service/pkg/storage"
 	"github.com/gitpod-io/gitpod/ws-sync/api"
 	"github.com/gitpod-io/gitpod/ws-sync/pkg/archive"
-	"github.com/gitpod-io/gitpod/ws-sync/pkg/cri"
 	"github.com/gitpod-io/gitpod/ws-sync/pkg/internal/session"
 	"github.com/gitpod-io/gitpod/ws-sync/pkg/quota"
 	"github.com/gitpod-io/gitpod/ws-sync/pkg/safetynet"
